Targeted refactor to useGigsWorkspace
